Remove static and single url builder
