+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Change Log
 
 
+## UNRELEASED
+
+### Removed
+
+- Single URL builder
+- Static URL builder
+
+
 ## 1.0.0-alpha - 2016-06-14
 
 - Initial alpha release
